requirements.txt
# Core ML/DL Libraries
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.10.0

# Optional for deployment
flask>=2.3.0
gradio>=3.35.0
streamlit>=1.24.0

Flask==3.0.0
ultralytics==8.1.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0